+++ consts.go
@@ -1,0 +1,14 @@
+package lightning
+
+// MIME types
+const (
+	MIMETextPlain       = "text/plain"
+	MIMEApplicationXML  = "application/xml"
+	MIMEApplicationJSON = "application/json"
+)
+
+// Header keys
+const (
+	HeaderContentType        = "Content-Type"
+	HeaderContentDisposition = "Content-Disposition"
+)