+++ examples/not_found/app.go
@@ -1,21 +1,17 @@
 package main
 
 import (
-	"fmt"
-	"github.com/go-labx/lightning"
 	"net/http"
+
+	"github.com/go-labx/lightning"
 )
 
 func main() {
-	app := lightning.NewApp()
-
-	app.NotFoundHandler = func(ctx *lightning.Context) {
-		ctx.Redirect(http.StatusMovedPermanently, "/404")
-		ctx.JSON(http.StatusOK, map[string]interface{}{
-			"code": 404,
-			"msg":  fmt.Sprintf("API %s -> %s not found.", ctx.Method, ctx.Path),
-		})
-	}
+	app := lightning.NewApp(&lightning.Config{
+		NotFoundHandler: func(ctx *lightning.Context) {
+			ctx.Text(404, "custom not found")
+		},
+	})
 
 	app.Get("/ping", func(ctx *lightning.Context) {
 		ctx.JSON(http.StatusOK, map[string]string{